Implement JWT-based authentication on the frontend
In this approach, the server will sign a token for the user using its
own private key. The user will send this token in each subsequent
request to prove the server has validated their login. This has the
benefit of being quicker than typical token-based authentication
because a database request is not required on each API call to verify
the users token.

Read more: https://jwt.io/introduction.