Get API URL from environment
The root URL of the API can be specified using the VUE_APP_ROOT_API
environment variable.  If not specified, the values in .env will be
used.

The README.md has also been updated to include instructions on how to
get a local development version of the frontend up and running.